requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
MouseInfo==0.1.3
Pillow==7.2.0
pushbullet.py==0.11.0
PyAutoGUI==0.9.50
PyGetWindow==0.0.8
PyMsgBox==1.0.8
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
python-magic==0.4.18
python3-xlib==0.15
PyTweening==1.0.3
requests==2.24.0
six==1.15.0
urllib3==1.25.10
websocket-client==0.57.0
